createpart3c.md
There is small confusion about errorHandler middleware. The documentation sates that errorHandler middleware shall be the last to be added  to use calls. I wrote app.use(errorHandler) to the end of use calls. However it did not work. The errorHandler was never called. I have verified with document several  times but of no use. Finally I have found from the code on the github, this call has to be made after all the route functions and it worked fine. 

The documentation need to be corrected accordingly.